New printout

+++ src/main/java/persistenceserver/Model/Group.java
@@ -2,15 +2,15 @@ package persistenceserver.Model;
 
 public class Group {
 
-    private int id;
+    private long id;
     private String name;
 
-    public Group(int id, String name) {
+    public Group(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
